Use autoloader for the Typogrify class.
Originally, two copies of the class file were included:
typogrify.class.php and src/Typogrify.php. The first of these was used
via module_load_include() and the second was ignored. The second was
tweaked in later commits.

This commit removes the first version of the file and the corresponding
module_load_include() statments, and modifies all references to the
Typogrify class to use the second (properly namespaced) version of the
class.